feat/fix(JIT compilation with Jitify):
Hide GPU architecture adaptions for the UDF LLMV from the unit tests and
move it into libcudf. This way, we ensure that the provided LLVM IR fits
to the current GPU architecture transparently for the user. In
particular, this avoids the necessity to change attributes for the
current target GPU architecture in unit tests
like the C++ unit tests or the hipDF JNI unit tests.