fix(cpp/.../statistics.cuh): Fix uninitialized variables of union type statistics_val,
Context: We have seen test failures of OrcStatisticsTest.Basic on MI300X.
Further analysis has revealed that when encoding the sum statistics,
an invalid value was written. This happened non-deterministically.
We traced this back to missing zero-initialization of *all* bytes
of the union `statistics_val sum` which is a member of statistics_chunk.
Re-ordering the members ensures correct zero-initialization.