Fix writing of compressed ORC files with large stripe footers (#17700)
In ORC, stripe footers can be compressed, the same way as the data. This means that compressed footers need to be written in multiple blocks if they are larger than the maximum block size. This applies even if the footer is actually uncompressed (in this case a flag in the block header is raised).
Currently, the ORC writer does not take into account that footer can be larger than max block size, and writes the entire thing in a single block, which is not valid.
The issue only applies to compressed files. Uncompressed files do not apply this limitation to the footers.

This PR changes the way the stripe footers are written to account for this case. The output hasn't changed for files with small stripe footers.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - https://github.com/nvdbaranec

URL: https://github.com/rapidsai/cudf/pull/17700