Remove pandas backend from `cudf.pandas` - ibis integration tests (#17945)
ibis has removed their pandas backend in version 10.0.0. In their
release notes,

> pandas: The pandas backend is removed. Note that pandas DataFrames are
STILL VALID INPUTS AND OUTPUTS and will remain so for the foreseeable
future. Please use one of the other local backends like DuckDB, Polars,
or DataFusion to perform operations directly on pandas DataFrames.

This PR removes the pandas backend from the integration tests. And
asserts that the inputs and outputs to ibis APIs are proxy objects.