Skip the failing connectorx polars tests (#18037)
In #18015, we tried skipping the failing polars tests and applying the
workaround mentioned in polars issue 21274. But pip is [unable to solve
our test
environment](https://github.com/rapidsai/cudf/actions/runs/13406947992/job/37463788766)
in that case. This PR just skips the tests because we only need to do
one or the other, not both.