fix(data_chunk_source_test.cpp): Synchronize copy operations correctly.
Details: With stream identification and PTDS, unit test failures have
been observed for DATA_CHUNK_SOURCE_TEST. They happen because a non-default stream is used which
is not synchronized with the per-thread default stream. In that case, it
could happen that the device data has not been written yet before it was
copied back to the host, leading to unit test failures.